requirements.txt
# Backend y API REST
Flask==3.0.0
flask-cors==4.0.0

# Procesamiento de datos
pandas>=2.1.4
numpy>=1.26.2

# Cálculos científicos y optimización
scipy>=1.11.4

# Base de datos
psycopg[binary]>=3.1.0

# Machine Learning (para integración con Diego)
scikit-learn>=1.3.2
xgboost>=2.0.3

# Visualización (para integración con Gianfranco)
plotly==5.18.0

# Chatbot (para integración con Ronaldo)
groq==0.4.1

# Utilidades
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0

# Email
Flask-Mail==0.9.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Producción
gunicorn==21.2.0

# Additional dependencies from feature branch
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
twilio
openpyxl
